refactor(voice): 인프라 패키지 위치 정리
