+++ cowork-voice/cmd/server/main.go
@@ -16,11 +16,11 @@ import (
 	mongoopts "go.mongodb.org/mongo-driver/v2/mongo/options"
 
 	"github.com/cowork/cowork-voice/internal/config"
-	"github.com/cowork/cowork-voice/internal/domain/channel"
-	kafkadomain "github.com/cowork/cowork-voice/internal/domain/kafka"
 	sessiondomain "github.com/cowork/cowork-voice/internal/domain/session"
 	webhookdomain "github.com/cowork/cowork-voice/internal/domain/webhook"
 	"github.com/cowork/cowork-voice/internal/health"
+	"github.com/cowork/cowork-voice/internal/infra/channel"
+	kafkadomain "github.com/cowork/cowork-voice/internal/infra/kafka"
 	"github.com/cowork/cowork-voice/internal/middleware"
 )
